[Fix] FacilityChecker 메서드 분리 및 Facility 열거형에 CaseIterable 적용
